move resnet into vision.models package
